ViewCounter entity created for tracking psot views

+++ src/main/java/com/topov/forum/dto/PostDto.java
@@ -13,7 +13,7 @@ public class PostDto {
     private String title;
     private String text;
     private String author;
-    private BigInteger views;
+    private Integer views;
     private Integer commentsAmount;
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
     private List<CommentDto> comments;
